requirements.txt
discord==1.0.1
discord.py==1.3.2
Pillow==7.0.0
nltk==3.4.5